requirements.txt
SpeechRecognition
PyAudio
psutil
pyttsx3
pyyaml
colorama
requests
pydub 
